+++ CHANGELOG.md
@@ -23,7 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- AAR038 - InfoTitleRequired: The `info.title` field must exist and not be empty.
+- AAR050 - InfoTitleRequired: The `info.title` field must exist and not be empty.
 
 ## [1.0.0] - 2025-12-29